Refactor the web routes and methods to a new package, pass in dependencies
